refactor: DispatcherModule common 모듈로 이동
