+++ core/common/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     alias(libs.plugins.noffice.android.library)
     alias(libs.plugins.noffice.android.library.compose)
     alias(libs.plugins.noffice.android.application.test)
+    alias(libs.plugins.noffice.android.hilt)
 }
 
 android {
