+++ core/common/src/main/java/com/easyhz/noffice/core/common/di/Qualifier.kt
@@ -1,0 +1,12 @@
+package com.easyhz.noffice.core.common.di
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Dispatcher(val dispatcher: NofficeDispatchers)
+
+enum class NofficeDispatchers {
+    DEFAULT,
+    IO,
+}
