--- data/announcement/src/main/java/com/easyhz/noffice/data/announcement/di/util/Qualifier.kt
@@ -1,7 +1,0 @@
-package com.easyhz.noffice.data.announcement.di.util
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class DispatcherIO
